FEAT: Install nextjs and make to be compilable
