+++ next.config.js
@@ -1,0 +1,14 @@
+// for transpiling all ESM @fullcalendar/* packages
+// also, for piping fullcalendar thru babel (to learn why, see babel.config.js)
+const withTM = require('next-transpile-modules')([
+  '@fullcalendar/common',
+  '@fullcalendar/daygrid',
+  '@fullcalendar/interaction',
+  '@fullcalendar/react',
+  '@fullcalendar/timegrid',
+])
+
+module.exports = withTM({
+  // any other general next.js settings
+  reactStrictMode: true,
+})
