+++ package.json
@@ -31,11 +31,12 @@
     "firebase": "^9.6.8",
     "graphql": "^16.3.0",
     "moment": "^2.29.1",
+    "next": "^12.1.0",
+    "next-transpile-modules": "^9.0.0",
     "react": "^17.0.2",
     "react-div-100vh": "^0.7.0",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.27.1",
-    "react-scripts": "5.0.0",
     "react-use": "^17.3.2",
     "typescript": "^4.5.5",
     "unsplash-js": "^7.0.15",
@@ -43,10 +44,9 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "generate": "graphql-codegen -r dotenv/config"
   },
   "eslintConfig": {
