+++ src/pages/_app.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { AppProps } from 'next/app'
+
+const App: React.FC<AppProps> = ({ Component, pageProps }) => {
+  return <Component {...pageProps} />
+}
+
+export default App
